[Justice Counts] Standalone errors/warnings page bug fix (Recidiviz/recidiviz-data#23109)
## Description of the change

Our first user successfully tested Automated Bulk Upload! This revealed
a small bug in the email confirmation. The email included a link to a
standalone errors/warnings page, however the uploaded file did not cause
any errors/warnings. This was due to an errant conditional. Before, we
were checking `if metric_key_to_errors is not None:`. However,
`metric_key_to_errors` will never be None because it is initialized as a
`defaultdict(list)`. To fix this, we change the conditional to check the
length of `metric_key_to_errors`.

I tested by re-deploying to nichelletest and copying the same uploaded
file to staging and it resulted in the following email (correctly does
not include a link to standalone errors/warnings page):
![Screenshot 2023-08-09 at 3 53 06
PM](https://github.com/Recidiviz/recidiviz-data/assets/82831800/7ac4ecc5-0153-4a77-97bd-0281ae3c3d72)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@michelles-mbp.lan>
GitOrigin-RevId: f1d174b643463adb609526768e7e4a62b88b4ea7