[StateStaff] Loosen invariant check to unblock ingest (Recidiviz/recidiviz-data#23159)
## Description of the change

There is a bug in how entity matching handles state_agents - if we
update a supervision period with external_id A so that it now does not
have a state_agent where before it did, we will clear out the
`supervising_officer_staff_external_id` field (correct), but not the
`supervising_officer_id` agent id field (incorrect). This is tripping up
our invariant checks which are mostly there to keep someone from
updating ingest to include state agents without hydrating
`supervising_officer_staff_external_id` at all. Since `state_agent` is
going to be deleted within the week, the pain this is causing for IEs is
worse than the risk of ingest changing between now and the `state_agent`
deprecation.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to
https://recidiviz.slack.com/archives/C01J3JV7962/p1690933415898039

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2e12828bae14a4684623e2e69697942c73fcc1b6