[Deploy] Add reminder about cherry-picks for when there are newly-added commits to the release branch (Recidiviz/recidiviz-data#23208)
## Description of the change

Adds the following prompt to `deploy_production.sh`:
```
There are newly-added commits in the release branch that are not in v1.543.0:
>  213f96741d [US_ND] Map supervision_site_id 19 & 22 to NULL (Recidiviz/recidiviz-data#23162) (Recidiviz/recidiviz-data#23195)
>  fa2f905c03 [CHERRYPICK][Admin Panel] Add additional condition for batches (Recidiviz/recidiviz-data#23183) (Recidiviz/recidiviz-data#23188)
This likely means folks expect these changes to go out with the next production deploy.
Would you like to continue deploying v1.543.0? (y/n):
```

prompt wording help appreciated!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7885f81bde2921d07a733d58e171848f08cabb0e