[US_CA] Change column name to unblock ingest (Recidiviz/recidiviz-data#23190)
## Description of the change

Starting on 8/14, CA `PersonParole` raw data started coming in with the
column formerly called `SUPVEVNTDATE` replaced with a column called
`LastCCRDate`. The `SUPVEVNTDATE` doesn't seem to be used in any ingest
views at this time, so I don't expect this change to have any impact
other than unblocking ingest in CA staging.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#23186

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c08ad3747c754ba4a77789af706df4dabebdd987