[US_PA][Outliers] Pad district external IDs with zeros (Recidiviz/recidiviz-data#23266)
## Description of the change

Since we do not have reliable location information for all officers in
PA, we are using the inferred locations in
`supervision_officer_attribute_sessions` (which are pulled from
`supervision_officer_inferred_location_sessions`) to determine the
location of some officers. This was causing a problem with outliers
email generation because those inferred location IDs are single digits,
not padded by zeros, while all the rest of our location external IDs are
padded. This change pads all the location external IDs pulled in from
the state staff query template with zeros up to two digits.

I tested these changes in a sandbox with prefix `elise_`. Queries I used
to validate that there are no single-digit district IDs are
[here](https://docs.google.com/document/d/1UmmuRs3WvCttIu0OwyOXV4mfhuyI3Rdj-zGITQCJTDg/edit?usp=sharing).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0dc23407213dac04259fe1fd81b466c00b5c3452