[US_MI] Revise sp normalization to drop bad open periods likely produced from lack of entity deletion (Recidiviz/recidiviz-data#23160)
## Description of the change

**Context:**
In MI, we'll often have a case where a person is discharge on a certain
date, and part of the raw data is closed out for that person on the
discharge date, but the movement raw data actually denoting the
discharge is entered retroactively a few days later. This leads to an
issue because in an earlier ingest run (on the date of the discharge),
we'll create an extra open period that starts on the discharge date
since we haven't observed that the person was discharged yet. Then when
the discharge movement is retroactively entered, the period that ends on
the discharge date will be appropriately updated with a
DISCHARGE/EXPIRATION termination reason, but we'll still be left with
this errant open period due to the lack of entity deletion. This leads
to clients showing up in Workflows when they were already properly
discharged.

We noted this issue before during workflows development, but because we
were frequently rerunning supervision periods for various revisions, the
issue self-resolved because MI is a state where we prune. But now that
workflows has launched and ingest revisions are less frequent, this
issue is growing increasingly large. Because reruns are time intensive
in MI, I'd like to handle as much as we can in normalization.

**Implementation:**
In this PR, I added a MI specific implementation of `drop_bad_periods`
(previously named `drop_bad_unmodified_periods`) to drop open periods
that:
- immediately succeed a period that ended with a DISCHARGE or
EXPIRATION,
  - don't start due to a new movement that's recorded, and
  - are the last supervision period for that person

This is a slightly conservative route, but I want to make sure this only
drops the periods we are sure are errant.

In this PR, I also remove the TN implementation of
`drop_bad_unmodified_periods` since it is no longer needed (verified
with @gheidkamp11 )

**Testing**
Loaded a sandbox in staging and tested to confirm that all periods we
would drop are periods that wouldn't be created if we ran the view on
only the _latest data (saved in `us_mi_sp_current_staging`):
```
WITH
n_periods_base_on_latest as (
  select offender_id, max(period_id) as total_real_periods
  from `recidiviz-staging.brenda_scratch.us_mi_sp_current_staging`
  group by 1
),
dropped_periods as (
  -- returns 706 rows
  select
    split(external_id, "-")[offset(0)] as offender_id,
    CAST(split(external_id, '-')[offset(1)] as int) as period_id
  from (
    select external_id
    from `recidiviz-staging.normalized_state.state_supervision_period`
    where state_code = 'US_MI'

    except distinct

    select external_id
    from `recidiviz-staging.brendali121_us_mi_normalized_state.state_supervision_period`
  )
)

-- returns no rows
select *
from dropped_periods
left join n_periods_base_on_latest USING(offender_id)
where period_id <= total_real_periods
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 18328eb8657143fdbec45b1d85a331857b99053b