[US_ME] Add completion event for furloughs (Recidiviz/recidiviz-data#23185)
## Description of the change

> Add completion event for furloughs based on the CIS_309_MOVEMENT
dataset. We are also going to receive information from another source,
so we'll compare them and decide which one to use once we get it.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7da67c40b9a075d9d06db3c3765cda4d284723fd