[US_MI] Do not group officers with unknown emails (Recidiviz/recidiviz-data#23281)
## Description of the change

The previous ingest view logic hydrating `state_staff` in MI grouped
officers by their email addresses to determine their names and ID
numbers. This doesn't work perfectly because 67 officers have the email
"no_email@michigan.gov", possibly because they no longer work in the
department, or they worked there before email was used, or some other
reason. Because all of these officers shared an email, they were being
ingested as one single person. This change makes it so that officers
with that email address are not grouped together, and are all ingested
as distinct staff members with one ID per person.

This was causing an invariant failure because as a side effect of this
grouping, an officer was being ingested with no external ID.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes https://github.com/Recidiviz/recidiviz-data/issues/23277

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c81fa6fa024d83cbe3386f515df8bf99fa578450